+++ packages/capacitor/src/plugin/PowerSyncCore.ts
@@ -1,6 +1,6 @@
 import { registerPlugin } from '@capacitor/core';
-import { PowerSyncPlugin } from './PowerSyncPlugin';
+import { PowerSyncPlugin } from './PowerSyncPlugin.js';
 
 export const PowerSyncCore = registerPlugin<PowerSyncPlugin>('PowerSync', {
-  web: () => import('./web').then((m) => new m.PowerSyncWeb())
+  web: () => import('./web.js').then((m) => new m.PowerSyncWeb())
 });