+++ packages/common/src/index.ts
@@ -32,6 +32,7 @@ export * from './db/schema/IndexedColumn.js';
 export * from './db/schema/Schema.js';
 export * from './db/schema/Table.js';
 export * from './db/schema/TableV2.js';
+export * from './db/schema/RawTable.js';
 
 export * from './client/Query.js';
 export { MemoryTriggerClaimManager } from './client/triggers/MemoryTriggerClaimManager.js';