+++ packages/web/vitest.config.ts
@@ -19,10 +19,8 @@ const config: UserConfigExport = {
        * they link to `.js` files.
        */
       '@powersync/web': path.resolve(__dirname, './lib/src'),
-      // https://jira.mongodb.org/browse/NODE-5773
-      bson: require.resolve('bson'),
       // Mock WebRemote to throw 401 errors for all HTTP requests in tests
-      '../../db/sync/WebRemote': path.resolve(__dirname, './tests/mocks/MockWebRemote.ts')
+      '../../db/sync/WebRemote.js': path.resolve(__dirname, './tests/mocks/MockWebRemote.ts')
     }
   },
   worker: {